refactor: 모델 파일을 한 파일-하나의 모델 구조로 분리
