+++ models/schemas/article.js
@@ -55,65 +55,4 @@ class Article extends Model {
   }
 }
 
-class Tag extends Model {
-  static init(sequelize) {
-    return super.init({
-      tag_id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-      },
-      name: {
-        type: DataTypes.STRING(50),
-        allowNull: false,
-        unique: true
-      }
-    }, {
-      sequelize,
-      tableName: 'tags',
-      timestamps: false
-    });
-  }
-
-  static associate(models) {
-    this.hasMany(models.ArticleTag, { foreignKey: 'tag_id', as: 'article_tags' });
-  }
-}
-
-class ArticleTag extends Model {
-  static init(sequelize) {
-    return super.init({
-      article_id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        references: {
-          model: 'articles',
-          key: 'article_id'
-        }
-      },
-      tag_id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        references: {
-          model: 'tags',
-          key: 'tag_id'
-        }
-      }
-    }, {
-      sequelize,
-      tableName: 'article_tags',
-      timestamps: false
-    });
-  }
-
-  static associate(models) {
-    this.belongsTo(models.Article, { foreignKey: 'article_id', as: 'article' });
-    this.belongsTo(models.Tag, { foreignKey: 'tag_id', as: 'tag' });
-  }
-}
-
-module.exports = {
-  Article,
-  Tag,
-  ArticleTag
-};
+module.exports = Article;
