+++ models/schemas/patient_hospital_link.js
