+++ models/schemas/question_ticket.js
