+++ models/schemas/refund.js
@@ -68,100 +68,4 @@ class Refund extends Model {
   }
 }
 
-class RefundTicket extends Model {
-  static init(sequelize) {
-    return super.init({
-      refund_ticket_id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-      },
-      refund_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: 'refunds',
-          key: 'refund_id'
-        }
-      },
-      user_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: 'users',
-          key: 'user_id'
-        }
-      },
-      ticket_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: 'question_tickets',
-          key: 'ticket_id'
-        }
-      },
-      created_at: {
-        type: DataTypes.DATE,
-        allowNull: false,
-        defaultValue: DataTypes.NOW
-      }
-    }, {
-      sequelize,
-      tableName: 'refund_tickets',
-      timestamps: false
-    });
-  }
-
-  static associate(models) {
-    this.belongsTo(models.Refund, { foreignKey: 'refund_id', as: 'refund' });
-    this.belongsTo(models.User, { foreignKey: 'user_id', as: 'user' });
-    this.belongsTo(models.QuestionTicket, { foreignKey: 'ticket_id', as: 'ticket' });
-  }
-}
-
-class AutoRefundSchedule extends Model {
-  static init(sequelize) {
-    return super.init({
-      schedule_id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-      },
-      question_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: 'questions',
-          key: 'question_id'
-        }
-      },
-      scheduled_at: {
-        type: DataTypes.DATE,
-        allowNull: false
-      },
-      executed_at: {
-        type: DataTypes.DATE,
-        allowNull: true
-      },
-      status: {
-        type: DataTypes.ENUM('PENDING', 'EXECUTED', 'CANCELLED'),
-        allowNull: false,
-        defaultValue: 'PENDING'
-      }
-    }, {
-      sequelize,
-      tableName: 'auto_refund_schedules',
-      timestamps: false
-    });
-  }
-
-  static associate(models) {
-    this.belongsTo(models.Question, { foreignKey: 'question_id', as: 'question' });
-  }
-}
-
-module.exports = {
-  Refund,
-  RefundTicket,
-  AutoRefundSchedule
-};
+module.exports = Refund;
