+++ server.js
@@ -10,7 +10,6 @@ const PORT = process.env.PORT;
 app.use(cors());
 app.use(express.json());
 
-// app.use('/api/v1', require('./routes'));
 
 app.listen(PORT, () => {
   console.log(`Server running on port ${PORT}`);
